requirements.txt
pandas>=1.5.0
numpy>=1.23.0
sqlalchemy>=2.0.0
pyyaml>=6.0
pytest>=7.0.0
python-dateutil>=2.8.0
openpyxl>=3.0.0
scipy>=1.9.0
